Updated eCommerce and UsersAPI REST homework

+++ UsersAPI/src/main/java/com/jeff/UsersAPI/controllers/UserController.java
@@ -3,7 +3,12 @@ package com.jeff.UsersAPI.controllers;
 import java.util.List;
 import java.util.Optional;
 
+import javax.validation.Valid;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -30,32 +35,67 @@ public class UserController {
 //	}
 	
 	@GetMapping("/users")
-	public List<User> getUsers(@RequestParam(value="state", required=false) String state){
+	public ResponseEntity<List<User>> getUsers(@RequestParam(value="state", required=false) String state){
+	List<User> users;
 	if (state != null) {
-	return (List<User>) userRepository.findByState(state);
+	//return (List<User>) userRepository.findByState(state);
+	users = userRepository.findByState(state);
+	return new ResponseEntity<List<User>> (users, HttpStatus.OK);
 	}
-	return (List<User>) userRepository.findAll();
+	//return (List<User>) userRepository.findAll();
+	users = userRepository.findAll();
+	return new ResponseEntity<List<User>> (users, HttpStatus.OK);
 	}
 	
+	//TODO: update this one to use response entity
 	@GetMapping("/users/{id}")
-	public Optional<User> getUserById(@PathVariable(value="id") Long id){
-	return userRepository.findById(id);
+	public ResponseEntity<Optional<User>> getUserById(@PathVariable(value="id") Long id, BindingResult bindingResult){
+		if (bindingResult.hasErrors()) {
+	        return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+	    }
+		Optional<User> foundUser = userRepository.findById(id);
+		
+		if (!foundUser.isPresent()) {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
+		return new ResponseEntity<Optional<User>>(foundUser, HttpStatus.OK);
+	
 	}
 	
 	@PostMapping("/users")
-	public void createUser(@RequestBody User user){
-	//vv this is how we save to the db.
+	public ResponseEntity<Void> createUser(@RequestBody @Valid User user,
+			BindingResult bindingResult){
+	if (bindingResult.hasErrors()) {
+        return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+    }
 	userRepository.save(user);
+	return new ResponseEntity<>(HttpStatus.CREATED);
 	}
 	
 	@PutMapping("/users/{id}")
-	public void createUser(@PathVariable(value="id") Long id, @RequestBody User user){
-	userRepository.save(user);
+	public ResponseEntity<Void> updateUser(@PathVariable(value="id") Long id, 
+			@RequestBody @Valid User user, BindingResult bindingResult){
+		if (bindingResult.hasErrors()) {
+	        return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+	    }
+		
+		Optional<User> foundUser = userRepository.findById(id);
+		if (!foundUser.isPresent()) {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
+		
+		userRepository.save(user);
+		return new ResponseEntity<>(HttpStatus.OK);
 	}
 	
 	@DeleteMapping("/users/{id}")
-	public void createUser(@PathVariable(value="id") Long id){
-	userRepository.deleteById(id);
+	public ResponseEntity<Void> deleteUser(@PathVariable(value="id") Long id){
+		Optional<User> foundUser = userRepository.findById(id);
+		if (!foundUser.isPresent()) {
+			return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+		}
+		userRepository.deleteById(id);
+		return new ResponseEntity<>(HttpStatus.OK);
 	}
 	
 }
